requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
django-jalali==7.2.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
jalali_core==1.0.0
jdatetime==5.0.0
mysqlclient==2.2.5
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
